+++ app/app.js
@@ -22,7 +22,6 @@ define([
     'features/chatbox/emoji',
     'features/chatbox/voiceChat',
     'features/modal/modalShowImageFull',
-    'features/modal/modalUpdateVersion',
     'features/notification/notification'
 
 ], (
@@ -49,7 +48,6 @@ define([
     emojiComp,
     voiceChatComp,
     modalShowImageFullComp,
-    modalUpdateVersionComp,
     notificationComp
 ) => {
     const {
@@ -106,20 +104,6 @@ define([
 
     const onAssignDataToStore = (data) => {
         setGeneral(data);
-    };
-
-    const onRegisterSW = () => {
-        if ('serviceWorker' in navigator && process.env.NODE_ENV === 'production') {
-            navigator.serviceWorker.addEventListener('controllerchange', () => {
-                modalUpdateVersionComp.onInit();
-            });
-
-            navigator.serviceWorker.register('sw.js').then(reg => {
-                setInterval(() => {
-                    reg.update();
-                }, 90000);
-            });
-        }
     };
 
     const onAssignAdvanceThemeBody = () => {
@@ -197,9 +181,9 @@ define([
     };
 
     const parseJwt = (token) => {
-        var base64Url = token.split('.')[1];
-        var base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-        var jsonPayload = decodeURIComponent(atob(base64).split('').map(function (c) {
+        const base64Url = token.split('.')[1];
+        const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+        const jsonPayload = decodeURIComponent(atob(base64).split('').map((c) => {
             return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
         }).join(''));
 
@@ -305,7 +289,6 @@ define([
         $notiBoard.addClass('run');
         languageComp.onInit();
         syncComp.onInitAgain();
-        onRegisterSW();
         onInitGeneralEvents();
         onAssignAdvanceThemeBody();