+++ app/shared/api.js
@@ -2,14 +2,18 @@ define([
     'shared/data',
     'app/constant',
     'shared/functions',
-    'features/logout/logout',
-    'axios'
+    'axios',
+    'features/modal/modalLogout',
+    'rxjs',
+    'rxjs/operators'
 ], (
     GLOBAL,
     constant,
     functions,
-    Logout,
-    axios
+    axios,
+    modalLogout,
+    Rx,
+    operators
 ) => {
     const {
         BASE_URL,
@@ -17,6 +21,14 @@ define([
         REFRESH_TOKEN,
         API_URL,
     } = constant;
+
+    const {
+        filter,
+        take,
+        switchMap
+    } = operators;
+
+    const refreshTokenSubject = new Rx.BehaviorSubject(null);
 
     const getHeaderJson = () => ({
         'Accept-Language': GLOBAL.getLanguage(),
@@ -50,15 +62,19 @@ define([
     };
 
     // Interceptors
-    axios.interceptors.request.use((request) => {
-        if (request.url.includes('xm/api/'))
-            request.headers['Authorization'] = `Bearer ${(functions.getDataToLocalApplication(ACCESS_TOKEN) || '')}`;
-        return request;
-    },
+    axios.interceptors.request.use(
+        (request) => {
+            if (request.url.includes('xm/api/') && functions.getDataToLocalApplication(ACCESS_TOKEN)) {
+                request.headers['Authorization'] = `Bearer ${functions.getDataToLocalApplication(ACCESS_TOKEN)}`;
+                request.headers['Accept-Language'] = GLOBAL.getLanguage();
+            }
+            return request;
+        },
         (error) => {
             return Promise.reject(error);
         });
 
+    let isRefreshing = false;
     axios.interceptors.response.use((response) => {
         GLOBAL.setNetworkStatus(true);
         return response.data;
@@ -67,27 +83,55 @@ define([
         console.log(originalConfig);
         if (error.response) {
             GLOBAL.setNetworkStatus(true);
-            if (error.response.status === 401 && !error.config.url.includes('/auth/') && !originalConfig._retry) {
-                originalConfig._retry = true;
+            if (error.config.url.includes('/auth/') || error.config.url.includes('/logout')) {
+                isRefreshing = false;
+                return Promise.reject(error);
+            } else if (error.response.status === 401) { // request has failed with 401 (token expired)
                 try {
-                    const response = await refreshToken();
-                    console.log(response);
-                    functions.setDataToLocalApplication(ACCESS_TOKEN, response.data.access_token);
-                    functions.setDataToLocalApplication(REFRESH_TOKEN, response.data.refresh_token);
-                    functions.setCookie(response.data.access_token, 3650);
+                    if (!isRefreshing) { // there is NOT a refreshing-token process in progress:
+                        isRefreshing = true;
+                        refreshTokenSubject.next(null);
 
-                    originalConfig.headers['Authorization'] = `Bearer ${response.data.token}`;
+                        // refresh the token in API and wait for response
+                        const response = await refreshToken();
 
-                    return axios(originalConfig);
+                        // after successfull response, store values on local storage
+                        functions.setDataToLocalApplication(ACCESS_TOKEN, response.data.access_token);
+                        functions.setDataToLocalApplication(REFRESH_TOKEN, response.data.refresh_token);
+                        functions.setCookie(response.data.access_token, 3650);
+
+                        // and send the response data to the subject
+                        // (data will be multicasted to observers subscribed to this subject)
+                        refreshTokenSubject.next(response.data);
+
+                        // refreshing process is finished
+                        isRefreshing = false;
+
+                        // retry the failed request
+                        return axios(originalConfig);
+                    } else { // there is a refreshing process in progress:
+                        // don't refresh token, instead suscribe to the 
+                        // refresh token subject and wait for a response
+                        const req = await refreshTokenSubject.pipe(
+                            filter(data => data != null), // wait for data to not be null
+                            take(1), // take the first value
+                            switchMap((data) => Rx.of(data)))
+                            .subscribe(() => originalConfig);
+
+                        // retry the failed request 
+                        return axios(req);
+                    }
                 } catch (_error) {
-                    Logout.cleanSession();
-                    if (_error.response && _error.response.data) {
-                        return Promise.reject(_error.response.data);
+                    isRefreshing = false;
+                    if (error.response) {
+                        modalLogout.onInit(_error.response.data?.details || 'Unexpected error while refreshing token.');
+                        return Promise.reject(new Error('Error refreshing token'));
                     }
                     return Promise.reject(_error);
                 }
+            } else {
+                return Promise.reject(error);
             }
-            return Promise.reject(error);
         }
         GLOBAL.setNetworkStatus(false);
         return Promise.reject(19940402);