+++ sw.js
@@ -29,8 +29,6 @@ const urls = [
 ];
 
 self.addEventListener('install', function (event) {
-    self.skipWaiting();
-
     event.waitUntil(
         caches.open(version)
             .then(function (cache) {
@@ -39,7 +37,7 @@ self.addEventListener('install', function (event) {
         )
     );
 });
-  
+
 self.addEventListener('activate', function (event) {
     event.waitUntil(caches.keys().then(function (keys) {
         return Promise.all(keys.filter(function (key) {
@@ -62,11 +60,13 @@ self.addEventListener('activate', function (event) {
 //         })
 //     );
 // });
-  
-self.addEventListener('fetch', function(event) {
-    var request = event.request;
 
-    if (request.method != 'GET' || (request.method == 'GET' && request.url.includes("xm/api/sync"))) return;
+self.addEventListener('fetch', function (event) {
+    var request = event.request;
+    if (request.method != 'GET' ||
+        (request.method == 'GET' && (request.url.includes("xm/api/sync") || request.url.includes("xm/oauth2")))) {
+        return;
+    }
 
     if (request.url.includes("xm/api")) {
         var finalResponse = fetch(request).then(function (response) {
@@ -75,6 +75,12 @@ self.addEventListener('fetch', function(event) {
     } else {
         event.respondWith(fromCache(event.request));
         event.waitUntil(update(event.request).then(refresh));
+    }
+});
+
+self.addEventListener('message', function (event) {
+    if (event.data.action === 'skipWaiting') {
+        self.skipWaiting();
     }
 });
 
@@ -98,7 +104,7 @@ function cacheRequest(request, response) {
     if (response.type === "error" || response.type === "opaque" || response.status != 200 || request.method != 'GET') {
         return Promise.resolve();
     }
-  
+
     return caches
         .open(version)
         .then(cache => cache.put(request, response.clone()));
@@ -107,23 +113,23 @@ function cacheRequest(request, response) {
 function refresh(response) {
     return self.clients.matchAll().then((clients) => {
         clients.forEach((client) => {
-        const message = {
-            type: 'refresh',
-            url: response.url,
-            eTag: response.headers.get('ETag')
-        };
-        client.postMessage(JSON.stringify(message));
+            const message = {
+                type: 'refresh',
+                url: response.url,
+                eTag: response.headers.get('ETag')
+            };
+            client.postMessage(JSON.stringify(message));
         });
     });
 }
-  
+
 const requestNotificationPermission = async () => {
     const permission = await Notification.requestPermission();
     // value of permission can be 'granted', 'default', 'denied'
     // granted: user has accepted the request
     // default: user has dismissed the notification permission popup by clicking on x
     // denied: user has denied the request.
-    if (permission !== 'granted'){
+    if (permission !== 'granted') {
         throw new Error('Permission not granted for Notification');
     }
 }